app/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
boto3==1.34.131
python-dotenv==1.0.1
psycopg2-binary==2.9.10
